+++ app/src/main/java/com/android/chordas/tracktrack/BARTService.java
@@ -1,9 +1,9 @@
 package com.android.chordas.tracktrack;
 
-import com.android.chordas.tracktrack.model.BARTModel;
+import com.android.chordas.tracktrack.model.BARTModelRoot;
 import retrofit2.Call;
 import retrofit2.http.GET;
-import retrofit2.http.Path;
+import retrofit2.http.Query;
 
 /**
  * Created by sam_chordas on 1/21/16.
@@ -11,9 +11,9 @@ import retrofit2.http.Path;
 public final class BARTService {
 
   public interface TrainAPI {
-    @GET("api/etd/{orig}/{key}")
-    Call<BARTModel> getTrains(
-        @Path("orig") String origin,
-        @Path("key") String key);
+    @GET("api/etd.aspx?cmd=etd")
+    Call<BARTModelRoot> getTrains(
+        @Query("orig") String origin,
+        @Query("key") String key);
   }
 }
